requirements.txt
google-cloud-texttospeech
openai
pyaudio
